==> ipynb-notebooks/1. Extract Json Data to CSV.ipynb <==
import requests
import json
import pandas as pd

# Get All Candidates

### Request Data

response = requests.get('https://interactive.guim.co.uk/2022/11/us-midterm-elections/results/production/results.json')
data_dict = response.json() # Format dictionary dari jsonnya

### Melihat Struktur Data

data_dict.keys()

print(data_dict['houseDetails'].keys())
data_dict['houseDetails']['AK00']

### Membuat list data kandidat

list_candidates = []
for item in data_dict['houseDetails'].keys():
  district_candidates = data_dict['houseDetails'][item]['details']['candidates'] #Contoh: [{'name': 'Michael Burgess', 'party': 'R', 'votes': 183379, 'isWinner': 'X', 'isIncumbent': True}, {'name': 'Mike Kolls', 'party': 'L', 'votes': 81208}]
  for candidate in district_candidates:
    candidate['district'] = item
    list_candidates.append(candidate)

list_candidates[:10]

### Mengubah list jadi dataframe pandas

candidates_df= pd.read_json(json.dumps(list_candidates))
candidates_df['isWinner'] = candidates_df['isWinner'].apply(lambda x: True if x=="X" else False)
candidates_df['isIncumbent'] = candidates_df['isIncumbent'].apply(lambda x: True if x==1 else False)
candidates_df = candidates_df[(candidates_df['party']=="R")|(candidates_df['party']=="D")]

candidates_df.groupby('party').count()

candidates_df[candidates_df['isWinner']==True]
#Kalo hasilnya dah keluar semua harusnya 435

### Daftar Nama Calon

print(candidates_df['name'].values)
print(len(candidates_df['name'].values))

### Menyimpan data ke csv

from google.colab import drive
drive.mount('/content/drive')
%cd '/content/drive/MyDrive/Kumpulan Tugas/Rekdat Akhir'

candidates_df.to_csv('candidates.csv',index=False)

# Get All House Election Winners per District (For Y Label)

response = requests.get('https://interactive.guim.co.uk/2022/11/us-midterm-elections/results/production/results.json')
data_dict = response.json() # Format dictionary dari jsonnya

house_winners_df = pd.DataFrame.from_dict(data_dict['houseWinners'],orient='index',columns=['Winners'])
house_winners_df["District"] = house_winners_df.index
house_winners_df = house_winners_df[["District","Winners"]]

house_winners_df

house_winners_df.to_csv("house_winners.csv",index=False)
