==> ipynb-notebooks/2b. Scraping Twitter (Twint).ipynb <==
!pip install snscrape

import snscrape
from tqdm import tqdm
import datetime as dt
import os
import snscrape.modules.twitter as sntwitter
import pandas as pd

def scrape_tweets(query, max_tweets=-1,output_path=None ): 
    
    
    tweets_list = []
    for i,tweet in enumerate(sntwitter.TwitterSearchScraper(query).get_items()):
        if max_tweets != -1 and i >= int(max_tweets):
            break
        tweets_list.append([tweet.date, tweet.id, tweet.content])
    df = pd.DataFrame(tweets_list, columns=['Datetime', 'Tweet Id', 'Text'])
    if type(output_path)==str:
      output_path = os.path.join(output_path,dt.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")+"-"+str(query)+".csv")
      df.to_csv(output_path, index=False)
      print("Data tweet tersimpan di",output_path)
    return df


df = scrape_tweets("tes",5)
df

### Scraping per kandidat

from google.colab import drive
drive.mount('/content/drive')
%cd '/content/drive/MyDrive/Kumpulan Tugas/Rekdat Akhir'

candidates_df = pd.read_csv('candidates.csv')

candidates_df = candidates_df[candidates_df['party'].apply(lambda x : x in ["D","R"])] #Filter hanya calon republikan dan demokrat saja

candidates_df

all_tweet_df = pd.DataFrame(columns=['Datetime', 'Tweet Id', 'Text','Candidate Name','Candidate Party','Candidate District'])
for index,row in tqdm(candidates_df.iterrows()):
  tweet_calon_ini_df = scrape_tweets(row['name'],max_tweets=200)
  tweet_calon_ini_df['Candidate Name'] = row['name']
  tweet_calon_ini_df['Candidate Party'] = row['party']
  tweet_calon_ini_df['Candidate District'] = row['district']
  all_tweet_df = pd.concat([all_tweet_df,tweet_calon_ini_df])

all_tweet_df.to_csv('tweet_200percalon.csv',index=False)
