==> ipynb-notebooks/4. Final Data Merge.ipynb <==
from google.colab import drive
drive.mount('/content/drive')
%cd '/content/drive/MyDrive/Kumpulan Tugas/Rekdat Akhir'

import pandas as pd

election_summary_df = pd.read_csv("house_winners.csv")
election_summary_df

candidates_df = pd.read_csv("candidates_detailed.csv")
candidates_df

sentiment_df = pd.read_csv("sentiment_per_district.csv")
sentiment_df = sentiment_df.drop("index",axis=1)
sentiment_df

election_summary_df = election_summary_df.merge(sentiment_df,on=["District"],how="left")
election_summary_df

poll_dataset_df = pd.read_csv("https://datawrapper.dwcdn.net/LkGXt/3/dataset.csv",storage_options={'User-Agent': 'Mozilla/5.0'})
poll_dataset_df["Name"] = poll_dataset_df["Name"].str.replace("-AL","00").str.replace("-","")
poll_dataset_df

election_summary_df = election_summary_df.merge(poll_dataset_df[["Name","Raw PVI","2020 Biden %","2020 Trump %","2016 Clinton %","2016 Trump %"]],how="left",left_on="District",right_on="Name")
election_summary_df = election_summary_df.drop("Name",axis=1)
election_summary_df

election_summary_df["D Popularity"] = election_summary_df["District"].apply(lambda x: candidates_df[(candidates_df["party"]=="D")&(candidates_df["district"]==x)]["googleNewsPopularity"].sum())
election_summary_df["R Popularity"] = election_summary_df["District"].apply(lambda x: candidates_df[(candidates_df["party"]=="R")&(candidates_df["district"]==x)]["googleNewsPopularity"].sum())

election_summary_df[["2020 Biden %","2020 Trump %","2016 Clinton %","2016 Trump %"]] = election_summary_df[["2020 Biden %","2020 Trump %","2016 Clinton %","2016 Trump %"]].applymap(lambda x : float(x.replace("%",""))/100)
election_summary_df

election_summary_df.columns

election_summary_df = election_summary_df[['District', 'D Positive Tweet', 'D Negative Tweet','R Positive Tweet', 
                                           'R Negative Tweet', 'Raw PVI', '2020 Biden %','2020 Trump %', '2016 Clinton %', 
                                           '2016 Trump %', 'D Popularity', 'R Popularity', 'Winners']]

election_summary_df.to_csv("election_summary.csv",index=False)
