==> ipynb-notebooks/5. XGBoost Classifier_Winner Predictor.ipynb <==
from google.colab import drive
drive.mount('/content/drive')
%cd '/content/drive/MyDrive/Kumpulan Tugas/Rekdat Akhir'

import xgboost as xgb
import pandas as pd

from sklearn import datasets
from sklearn import metrics
from sklearn.model_selection import train_test_split

df = pd.read_csv("election_summary.csv")
df = df.dropna()

df

X = df.drop(["District","Winners"],axis=1)
y= df["Winners"]
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)

model = xgb.XGBClassifier()
model.fit(X_train, y_train)

print(metrics.classification_report(y_test, model.predict(X_test)))
print(metrics.confusion_matrix(y_test, model.predict(X_test)))

pd.DataFrame.from_dict(model.get_booster().get_score(importance_type='gain'),orient='index').sort_values(0,ascending=False)

pd.DataFrame.from_dict(model.get_booster().get_score(importance_type='weight'),orient='index').sort_values(0,ascending=False)

model.save_model("us_election_xgb_model.json")
